Changed to 'secret' virtualenv location
This is due to the fact that when cron runs, the current working
directory is not where the script is located (as I wrongfully thought)
but it's /root. So it will try to find the script in /root/venv/etc
which is wrong. I changed this to a 'secret' variable rather than
dynamically fetching it.